# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Fully fuctional dating app API",
        "title": "Love Bites API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081/user",
    "paths": {
        "/details": {
            "post": {
                "description": "Api for send user basic detials",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AuthSvc"
                ],
                "summary": "Api for send user details",
                "operationId": "UserDetails",
                "parameters": [
                    {
                        "description": "User details",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserDetails"
                        }
                    },
                    {
                        "type": "string",
                        "default": "Bearer",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User detials added succesfully",
                        "schema": {
                            "$ref": "#/definitions/responce.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/responce.Response"
                        }
                    }
                }
            }
        },
        "/sendotp": {
            "post": {
                "description": "Api for send otp to user phone",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AuthSvc"
                ],
                "summary": "Api for send otp",
                "operationId": "SendOtp",
                "parameters": [
                    {
                        "description": "Phone number",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.OtpReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Otp send succesfully",
                        "schema": {
                            "$ref": "#/definitions/responce.Response"
                        }
                    },
                    "400": {
                        "description": "Failed to send otp",
                        "schema": {
                            "$ref": "#/definitions/responce.Response"
                        }
                    }
                }
            }
        },
        "/upload/photos": {
            "post": {
                "description": "Api for upload user photos max 4 photos min 2 photos",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MatchSvc"
                ],
                "summary": "Api for upload photos",
                "operationId": "UploadPhotos",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Value for the photos field",
                        "name": "photos",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Bearer",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Phoptos added succesfully",
                        "schema": {
                            "$ref": "#/definitions/responce.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/responce.Response"
                        }
                    }
                }
            }
        },
        "/verifyotp": {
            "post": {
                "description": "Api for verify otp and authentication it will return token and user found or not",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AuthSvc"
                ],
                "summary": "Api for verify otp and auth",
                "operationId": "VerifyOtpAndAuth",
                "parameters": [
                    {
                        "description": "OTP and Phone number",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.VerifyOtpReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verifeid succesfully",
                        "schema": {
                            "$ref": "#/definitions/responce.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/responce.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.OtpReq": {
            "type": "object",
            "properties": {
                "phone": {
                    "type": "string"
                }
            }
        },
        "request.UserDetails": {
            "type": "object",
            "required": [
                "dob",
                "email",
                "fullname",
                "gender",
                "location"
            ],
            "properties": {
                "dob": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                }
            }
        },
        "request.VerifyOtpReq": {
            "type": "object",
            "required": [
                "otp",
                "phone"
            ],
            "properties": {
                "otp": {
                    "type": "string",
                    "minLength": 6
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "responce.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}